Revert "Refactored code to match team format"
