+++ client/src/components/Trade/Trade.jsx
@@ -14,7 +14,7 @@ function Trade({authenticatedUser, portfolio}) {
         </div>
         <Order authenticatedUser={authenticatedUser} portfolio={portfolio}/>
       </div>
-      <Newsfeed symbol={"BTC"}/>
+      <Newsfeed/>
     </div>
   )
 }
