+++ db/init.sql
@@ -81,21 +81,6 @@ CREATE TABLE IF NOT EXISTS leaderboard (
   UNIQUE (trader_id, coin_id)
 );
 
-CREATE TABLE IF NOT EXISTS newsfeed (
-  id SERIAL PRIMARY KEY,
-  title VARCHAR(200) NOT NULL,
-  description TEXT NOT NULL,
-  arthur VARCHAR(200) NOT NULL,
-  URL VARCHAR(200) NOT NULL,
-  tickers TEXT [] NOT NULL,
-  image_url VARCHAR(200) NOT NULL,
-  publish_date timestamp with time zone,
-  topics TEXT,
-  type TEXT,
-  sentiment TEXT
-);
-
-
 -- CREATE RULE update_leaderboard_on_insert_transactions AS ON INSERT TO transactions DO ALSO
 --   (
 --   INSERT INTO leaderboard
