+++ models/models.js
@@ -4,7 +4,6 @@ const Trade = require('./trade/trade.js');
 const Leaderboard = require('./Leaderboard/leaderboard.js');
 const Achievements = require('./achievements/Achievements.js');
 const Market = require('./marketWatch/marketWatch.js');
-const Newsfeed = require('./newsfeed/newsfeed.js');
 
 module.exports = {
   Auth,
@@ -12,6 +11,5 @@ module.exports = {
   Trade,
   Leaderboard,
   Achievements,
-  Newsfeed,
   Market
 }
